Adapt examples
