+++ examples/dummy_hook.rs
